feat: add doc window to completion and refactor code
